# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FairChat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get a user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        publicKey:
          type: string
        createdAt:
          type: string
          format: date-time
        isOnline:
          type: boolean
      required:
        - id
        - username
        - displayName
        - publicKey
        - createdAt
        - isOnline
...